requirements.txt
aiohttp
aiohttp-retry
aiosignal
anyio
attrs
certifi
charset-normalizer
click
distro
firebase-admin
Flask
flask-session
frozenlist
google-cloud-firestore
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
h11
httpcore
httpx
idna
itsdangerous
Jinja2
jiter
levenshtein
MarkupSafe
multidict
openai
pydantic
PyJWT
python-dateutil
python-dotenv
pytz
rapidfuzz
requests
schedule
sniffio
tqdm
twilio
typing_extensions
urllib3
Werkzeug
yarl